dead end, shibboleth login for email link
The email link redirects to a shibboleth login. I missed this when I
was going over things as it had stayed in my browser, and so
even though it showed me as logged out, it had completed
a login outside that to get to the links content. There is no way to
automate this unless there is an API with Pivot.